+++ NEWS.md
@@ -4,7 +4,8 @@
   * Move check_template to create_final
   * Move check_lint to create_final
   * Try to detect programming lanaguage
-  * 
+  * Ensure library calls use quotes
+  * Check margin notes & captions finish with a full stop
 
 # 0.5.0
   * Extract latex tags using xparse from latex